seismicupload/requirements.txt
azure-identity
azure-keyvault-secrets
azure-storage-blob
cerberus
openvds
psycopg[binary]
pyyaml
